requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-dotenv
google-generativeai
PyMuPDF
Pillow
pandas
python-jose[cryptography]
passlib[bcrypt]